Remove need for EditableComboBoxHighlightTextBlock object.
