+++ samples/SampleApp/DemoPages/EditableComboBoxDemo.axaml
@@ -275,7 +275,7 @@
                 <Border DockPanel.Dock="Left" Background="Gray" CornerRadius="9" Width="18" Height="18" Padding="2">
                   <TextBlock VerticalAlignment="Center" HorizontalAlignment="Center" Text="{Binding Code}" Foreground="White" FontWeight="Bold" FontSize="8" />
                 </Border>
-                <EditableComboBoxHighlightTextBlock
+                <SearchHighlightTextBlock
                   Content="{Binding Name}"
                   VerticalAlignment="Center"
                   Margin="6 0 0 0" />
