+++ src/Devolutions.AvaloniaControls/Controls/EditableComboBox/EditableComboBox.axaml
@@ -82,7 +82,7 @@
     <Setter Property="Height" Value="{DynamicResource EditableComboBoxItemsHeight}" />
     <Setter Property="ContentTemplate">
       <DataTemplate>
-        <EditableComboBoxHighlightTextBlock />
+        <SearchHighlightTextBlock />
       </DataTemplate>
     </Setter>
     <Setter Property="Template">
