+++ src/Devolutions.AvaloniaTheme.MacOS/Controls/EditableComboBox.axaml
@@ -92,7 +92,7 @@
     <Setter Property="Padding" Value="{DynamicResource ComboBoxItemPadding}" />
     <Setter Property="ContentTemplate">
       <DataTemplate>
-        <EditableComboBoxHighlightTextBlock />
+        <SearchHighlightTextBlock />
       </DataTemplate>
     </Setter>
     <Setter Property="Template">
